--- a/lib/logflare/source/bigquery/pipeline.ex
+++ b/lib/logflare/source/bigquery/pipeline.ex
@@ -36,7 +36,11 @@ defmodule Logflare.Source.BigQuery.Pipeline do
             default: [concurrency: System.schedulers_online() * 2]
           ],
           batchers: [
-            bq: [concurrency: System.schedulers_online() * 2, batch_size: 250, batch_timeout: 1_500]
+            bq: [
+              concurrency: System.schedulers_online() * 2,
+              batch_size: 250,
+              batch_timeout: 1_500
+            ]
           ],
           context: rls
         ],